FEniCS/Firedrake backends: Move interface code
