+++ tests/firedrake/test_equations.py
@@ -1,7 +1,7 @@
 from firedrake import *
 from tlm_adjoint.firedrake import *
+from tlm_adjoint.firedrake.backend_interface import assemble_linear_solver
 from tlm_adjoint.firedrake.expr import extract_coefficients
-from tlm_adjoint.firedrake.solve import assemble_linear_solver
 
 from .test_base import *
