+++ tlm_adjoint/fenics/caches.py
@@ -11,7 +11,8 @@ from ..interface import (
 
 from ..caches import Cache
 
-from .assembly import assemble, assemble_matrix, matrix_copy
+from .backend_interface import (
+    assemble, assemble_matrix, linear_solver, matrix_copy)
 from .expr import (
     derivative, eliminate_zeros, expr_zero, extract_coefficients, form_cached,
     replaced_form)
@@ -433,7 +434,6 @@ class LinearSolverCache(Cache):
             assembly_cache = globals()["assembly_cache"]()
 
         def value():
-            from .solve import linear_solver
             _, (A, b_bc) = assembly_cache.assemble(
                 form, bcs=bcs,
                 form_compiler_parameters=form_compiler_parameters,
